pyproject.toml
[tool.ruff]

line-length = 100
exclude = [".git", ".venv", "venv", "__pycache__", "build", "dist", ".env", ".cache", "poetry.lock"]


[tool.ruff.lint]

select = ["E", "W", "F", "I", "B", "C90", "UP"]
ignore = ["D100", "D104", "B008"]


[tool.ruff.lint.isort]

combine-as-imports = true
lines-after-imports = 2


[tool.ruff.format]

quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false


[tool.pytest.ini_options]

testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra -q --doctest-modules"


[tool.mypy]

python_version = "3.12"
strict = true
ignore_missing_imports = true
[project]
name = "project-root"
version = "0.1.0"
description = ""
authors = [
    {name = "Kostya",email = "fesenko.kostya576@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.8,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
